Finish creation dialog
Right now you can't paint the eyes or anything. And it just
updates the current daruma. I will soon make it so you can
have a list of them and move through it like a carrousel.
Or maybe a grid would be better. After all, I want them to
be all visible when I open the page.