Move drawing to dialog and extract overlapping
The overlapping was a bit weird inside the dialog,
so I made a custom OverlapComponents component that
just puts all children on top of each other. The idea
is this can be configured in the future to be left/right
part of the image, so it can be re-used in the goal finish.